refactor region checks
