+++ cmd/regioncheck/dreamhost.go
@@ -1,1 +1,5 @@
 package main
+
+func GetRegionsDreamhost() ([]string, error) {
+	return []string{"us-east-1"}, nil
+}
